🚚 [projects] Move constants `LATEST_BRANCH` and `UPDATE_BRANCH` from `services` to `create` module
