+++ src/cutty/projects/create.py
@@ -5,4 +5,8 @@ from pathlib import Path
 from cutty.repositories.domain.repository import Repository as Template
 
 
+LATEST_BRANCH = "cutty/latest"
+UPDATE_BRANCH = "cutty/update"
+
+
 CreateProject = Callable[[Path], Template]
