+++ src/cutty/services/update.py
@@ -5,10 +5,10 @@ from pathlib import PurePosixPath
 from typing import Optional
 
 from cutty.projects.create import CreateProject
+from cutty.projects.create import LATEST_BRANCH
+from cutty.projects.create import UPDATE_BRANCH
 from cutty.repositories.domain.repository import Repository as Template
 from cutty.services.create import create
-from cutty.services.create import LATEST_BRANCH
-from cutty.services.create import UPDATE_BRANCH
 from cutty.templates.adapters.cookiecutter.projectconfig import readprojectconfigfile
 from cutty.templates.domain.bindings import Binding
 from cutty.util.git import Repository
