+++ tests/unit/services/test_link.py
@@ -7,9 +7,9 @@ import pytest
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path as VirtualPath
 from cutty.projects.create import CreateProject
+from cutty.projects.create import LATEST_BRANCH
+from cutty.projects.create import UPDATE_BRANCH
 from cutty.repositories.domain.repository import Repository as Template
-from cutty.services.create import LATEST_BRANCH
-from cutty.services.create import UPDATE_BRANCH
 from cutty.services.link import linkproject
 from cutty.util.git import Repository
 from tests.util.git import updatefile
